moved main.py to project root
